assignment2/requirements.txt
flask
google-cloud-storage
google-cloud-firestore
google-auth 
google-auth-oauthlib
google-auth-httplib2
gunicorn